% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.flickr.R
\name{authFlickr}
\alias{authFlickr}
\title{Authentification function for Flickr API}
\usage{
authFlickr()
}
\description{
This function allows user identification  
Opens the browser and you can enter Flickr account details
}
