% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadImages.R
\name{downloadImages}
\alias{downloadImages}
\title{Download images found using \code{photosSearch}}
\usage{
downloadImages(photoSearch_results, licenses = 7:10, saveDir = ".",
  max_quality = 2, verbose = TRUE)
}
\arguments{
\item{photoSearch_results}{A data.frame output from \code{photoSearch}.}

\item{licenses}{Numeric vector, set the licenses you want to download images for. See \code{getLicenses}}

\item{saveDir}{Character, the path where images should be saved}

\item{max_quality}{Numeric 1-4 giving the maximum quality of image you want to download 1=small, 4=large}

\item{verbose}{logical, if TRUE the progeress through images is given}
}
\value{
A vector of all the URLs that where targetted for download
}
\description{
Using the data returned from the
}
\examples{
# run a workflow, using the logistic regression model
\dontrun{

snowmen <- photosSearch(year_range = c(2015, 2016),
                        text = 'snowman',
                        woe_id = 12578048)
downloadImages(snowmen,
               licenses = c(6:10),
               saveDir = tempdir(),
               max_quality = 2)

}
}
