% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPlaces.R
\name{findPlaces}
\alias{findPlaces}
\title{Find correct woe_id for photosSearch query}
\usage{
findPlaces(place)
}
\arguments{
\item{place}{Text string describing the place for the query}
}
\value{
A dataframe of places incl wo_id to be used in the photosSearch function
}
\description{
Find correct woe_id for photosSearch query
}
