% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLicenses.R
\name{getLicenses}
\alias{getLicenses}
\title{Get a lookup table of licenses}
\usage{
getLicenses()
}
\value{
a dataframe of license definitions
}
\description{
Get a lookup table of licenses
}
\examples{
# run a workflow, using the logistic regression model
\dontrun{

getLicenses()

}
}
