% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosSearch.R
\name{photosSearch}
\alias{photosSearch}
\title{Query Flickr API and build a dataframe of metadata}
\usage{
photosSearch(year_range, text, woe_id, has_geo = TRUE)
}
\arguments{
\item{year_range}{The year to search between (inclusive). For example c(2005,2007) will search for images from 2005, 2006 and 2007}

\item{text}{Set keywords so that query returns pictures with the text (e.g. "bird")}

\item{woe_id}{A 32-bit identifier that uniquely represents spatial entities (e.g. 12578048 is Scotland).}

\item{has_geo}{Logical, should the function only return records that have spatial data.}
}
\value{
A dataframe of metadata
}
\description{
This function uses the Flickr API to get metadata about the images
}
\examples{
# run a workflow, using the logistic regression model
\dontrun{

photosSearch(c(2005,2006), "bird", "12578048", TRUE)

}
}
