% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadImages.R
\name{downloadImages}
\alias{downloadImages}
\title{Download images found using \code{photosSearch}}
\usage{
downloadImages(photoSearch_results, licenses = 7:10, saveDir = ".")
}
\arguments{
\item{photoSearch_results}{A data.frame output from \code{photoSearch}.}

\item{licenses}{Numeric vector, set the licenses you want to download images for}

\item{saveDir}{Character, the path where images should be saved}
}
\description{
Using the data returned from the
}
\examples{
# run a workflow, using the logistic regression model
\dontrun{

snowmen <- photosSearch(year_range = c(2005, 2016),
                        text = 'snowman',
                        woe_id = 12578048)
downloadImages(images, licenses = c(4, 7:10))

}
}
